Set minimum macos version to 10.11 to fix Cocoapods Xcode 14.3+ issue
When Xcode 14.3 came out, they've removed a library called libarclitemacos.a from the
installation. This means that cocoapods builds with macos targets < 10.11 will fail, as
it will attempt to link against this missing library.

See the bug here :-  https://github.com/CocoaPods/CocoaPods/issues/11839

There has been discussion regarding implementing flags in cocoapods to work around the
issue but this has yet to materialize.

It appears that the only possible solution at this point is to increase the minimum
macos version in a podspec to 10.11 to avoid linking against libarclite_macosx.a.

The current ios, tvos and watchos minimum deployment targets are fine.

If we mark this change as v2.0.0, people who still require use of macos 10.10
(and older versions of Xcode) via Cocoapods can still link against the current v1.4.0.